chore(models): replace new latent module
