+++ src/exp/models/latent_policy.py
@@ -1,300 +1,224 @@
-"""Policy construct with Latent Encoder and Predictor."""
+from __future__ import annotations
 
+from collections.abc import Callable
 from typing import override
 
 import torch
 import torch.nn as nn
-from pamiq_core.torch import TorchTrainingModel
 from torch import Tensor
 from torch.distributions import Distribution
 
-from .components.deterministic_normal import FCDeterministicNormalHead
-from .components.multi_discretes import MultiEmbeddings
-from .components.stacked_features import LerpStackedFeatures, ToStackedFeatures
-from .components.stacked_hidden_state import StackedHiddenState
+from .components import (
+    FCDeterministicNormalHead,
+    FCMultiCategoricalHead,
+    FCScalarHead,
+    LerpStackedFeatures,
+    StackedHiddenState,
+)
+from .policy import HiddenStatePiV
 from .utils import ActionInfo, ObsInfo
 
 
-class LatentPolicy(nn.Module):
-    """Policy model that outputs latent representations and predicts next
-    observations from those representations.
+class LatentPiVFramework(HiddenStatePiV):
+    """Policy-Value framework with separated Encoder and Generator components.
 
-    This model combines an encoder that processes observations and
-    actions into latent representations, and a predictor that generates
-    the distribution of next observations based on those latent
-    representations.
+    Attributes:
+        encoder: Module that encodes observations into latent representations.
+        generator: Module that generates policy distributions and value estimates from latent representations.
     """
 
-    @override
     def __init__(
         self,
-        obs_upper_action_flatten_head: nn.Module,
-        encoder: StackedHiddenState,
-        predictor: StackedHiddenState,
-        action_dist_head: nn.Module,
-        value_head: nn.Module,
+        encoder: Encoder,
+        generator: Generator,
     ) -> None:
-        """Initialize the LatentPolicy model.
+        """Initialize the LatentPiVFramework.
 
         Args:
-            obs_action_flatten_head: The head that processes observations and actions
-            encoder: The stacked hidden state model for encoding observation-action pairs.
-            predictor: The stacked hidden state model for predicting next observations.
-            obs_predict_head: The head that generates the distribution of next observations
-                from latent representations.
+            encoder: Encoder module for processing observations.
+            generator: Generator module for producing policy and value outputs.
         """
         super().__init__()
-        self.obs_upper_action_flatten_head = obs_upper_action_flatten_head
+
         self.encoder = encoder
-        self.predictor = predictor
-        self.action_dist_head = action_dist_head
-        self.value_head = value_head
+        self.generator = generator
+
+    @override
+    def forward(
+        self,
+        observation: Tensor,
+        hidden: Tensor | None = None,
+        upper_action: Tensor | None = None,
+        *,
+        no_len: bool = False,
+    ) -> tuple[Distribution, Tensor, Tensor]:
+        """Forward pass through the framework.
+
+        Args:
+            obs: Observation tensor. Shape depends on encoder configuration.
+            hidden: Optional hidden state for the encoder. Shape is (*batch, depth, dim).
+
+        Returns:
+            A tuple containing:
+                - Policy distribution from the generator.
+                - Value estimate from the generator.
+                - Updated hidden state from the encoder.
+        """
+        x, hidden = self.encoder(observation, hidden, upper_action, no_len=no_len)
+        policy_dist, value = self.generator(x)
+        return policy_dist, value, hidden
+
+
+class Encoder(nn.Module):
+    """Encoder module for processing observations into latent
+    representations."""
+
+    def __init__(
+        self,
+        obs_info: ObsInfo | int,
+        core_model_dim: int,
+        core_model: StackedHiddenState,
+        out_dim: int | None = None,
+        upper_action_dim: int | None = None,
+    ) -> None:
+        """Initialize the Encoder.
+
+        Args:
+            obs_info: Observation configuration. If ObsInfo, uses lerp-based feature
+                extraction. If int, treats as direct observation dimension.
+            core_model_dim: Input dimension for the core model.
+            core_model: Stacked hidden state model for processing features.
+            out_dim: Optional output dimension. If None, uses core_model output dimension.
+            upper_action_dim: Optional dimension for upper-level policy actions. If provided,
+                enables hierarchical policy by concatenating upper-level actions with observations.
+        """
+        super().__init__()
+
+        # Setup observation projection.
+        self.obs_flatten = nn.Identity()
+        obs_action_dim = 0
+        if isinstance(obs_info, ObsInfo):
+            self.obs_flatten = LerpStackedFeatures(
+                obs_info.dim, obs_info.dim_hidden, obs_info.num_tokens
+            )
+            obs_action_dim += obs_info.dim_hidden
+        else:
+            obs_action_dim += obs_info
+
+        # Setup upper action projection
+        if upper_action_dim is not None:
+            obs_action_dim += upper_action_dim
+
+        self.obs_action_proj = nn.Linear(obs_action_dim, core_model_dim)
+        self.core_model = core_model
+
+        self.out_proj = nn.Identity()
+        if out_dim is not None:
+            self.out_proj = nn.Linear(core_model_dim, out_dim)
 
     @override
     def forward(
         self,
         obs: Tensor,
-        upper_action: Tensor,
-        hidden_encoder: Tensor | None = None,
-        hidden_predictor: Tensor | None = None,
+        hidden: Tensor | None = None,
+        upper_action: Tensor | None = None,
+        *,
         no_len: bool = False,
-    ) -> tuple[Distribution, Tensor, Tensor, Tensor, Tensor]:
-        """Process observation and upper action, compute latent representation,
-        and predict next observation distribution.
+    ) -> tuple[Tensor, Tensor]:
+        """Encode observations into latent representations.
 
         Args:
-            obs: Input observation tensor (*batch, len, num_token, obs_dim).
-            upper_action: Upper action tensor (*batch, len, num_token, action_dim).
-            hidden_encoder: Optional hidden state for the encoder (*batch, depth, dim).
-            hidden_predictor: Optional hidden state for the predictor (*batch, depth, dim).
-            no_len: If True, ignores sequence length dimension.
+            obs: Observation tensor. Shape is (*batch, len, num_tokens, obs_dim) if ObsInfo
+                was provided, otherwise (*batch, len, obs_dim).
+            hidden: Optional hidden state from previous timestep. Shape is (*batch, depth, dim).
+                If None, the core model will initialize its hidden state.
+            upper_action: Optional upper-level policy action tensor for hierarchical policies.
+                Shape is (*batch, len, upper_action_dim). If provided, concatenated with
+                observations before processing.
+            no_len: If True, processes inputs as single-step without sequence length dimension.
 
         Returns:
             A tuple containing:
-                - Distribution representing predicted observations.
-                - Value estimates from the value head.
-                - Latent representation from the encoder.
-                - Next hidden state for the predictor.
-                - Next hidden state for the encoder.
+                - Latent representation tensor of shape (*batch, len, out_dim).
+                - Updated hidden state tensor of shape (*batch, depth, dim).
         """
-        latent = self.obs_upper_action_flatten_head(obs, upper_action)
-
-        latent_encoder, next_hidden_encoder = self.encoder(
-            latent, hidden_encoder, no_len=no_len
-        )
-        latent_predictor, next_hidden_predictor = self.predictor(
-            latent_encoder, hidden_predictor, no_len=no_len
-        )
-
-        action_dist = self.action_dist_head(latent_predictor)
-        value = self.value_head(latent_predictor)
-
-        return (
-            action_dist,
-            value,
-            latent_encoder,
-            next_hidden_predictor,
-            next_hidden_encoder,
-        )
+        x = self.obs_flatten(obs)
+        if upper_action is not None:
+            x = torch.cat([x, upper_action], dim=-1)
+        x = self.obs_action_proj(x)
+        x, hidden = self.core_model.forward(x, hidden, no_len=no_len)
+        return self.out_proj(x), hidden
 
     @override
     def __call__(
         self,
         obs: Tensor,
-        upper_action: Tensor,
-        hidden_encoder: Tensor | None = None,
-        hidden_predictor: Tensor | None = None,
+        hidden: Tensor | None = None,
+        upper_action: Tensor | None = None,
+        *,
         no_len: bool = False,
-    ) -> tuple[Distribution, Tensor, Tensor, Tensor, Tensor]:
+    ) -> tuple[Tensor, Tensor]:
         """Override __call__ with proper type annotations.
 
         See forward() method for full documentation.
         """
-        return super().__call__(
-            obs, upper_action, hidden_encoder, hidden_predictor, no_len=no_len
-        )
+        return super().__call__(obs, hidden, upper_action, no_len=no_len)
 
-    def forward_with_no_len(
+
+class Generator(nn.Module):
+    """Generator module for producing policy distributions and value estimates
+    from latent representations."""
+
+    def __init__(
         self,
-        obs: Tensor,
-        upper_action: Tensor,
-        hidden_encoder: Tensor | None = None,
-        hidden_predictor: Tensor | None = None,
-    ) -> tuple[Distribution, Tensor, Tensor, Tensor, Tensor]:
-        """Forward with data which has no len dimension.
+        latent_dim: int,
+        action_info: ActionInfo | int,
+        core_model: nn.Module | None = None,
+        core_model_dim: int | None = None,
+    ) -> None:
+        """Initialize the Generator.
 
         Args:
-            obs: Input observation tensor (*batch, num_token, obs_dim).
-            upper_action: Upper action tensor (*batch, num_token, action_dim).
-            hidden_encoder: Optional hidden state for the encoder (*batch, depth, dim).
-            hidden_predictor: Optional hidden state for the predictor (*batch, depth, dim).
+            latent_dim: Dimension of input latent representations.
+            action_info: Action configuration specifying action structure.
+            core_model: Optional core model for processing latents. If None,
+                uses identity transformation.
+            core_model_dim: Hidden dimension for core model. If provided,
+                adds input projection from latent_dim to core_model_dim.
+        """
+        super().__init__()
+
+        self.input_proj = nn.Identity()
+        if core_model_dim is not None:
+            self.input_proj = nn.Linear(latent_dim, core_model_dim)
+
+        if core_model is None:
+            core_model = nn.Identity()
+        self.core_model = core_model
+
+        head_dim_in = latent_dim if core_model_dim is None else core_model_dim
+        if isinstance(action_info, ActionInfo):
+            self.policy_head = FCMultiCategoricalHead(head_dim_in, action_info.choices)
+        else:
+            self.policy_head = FCDeterministicNormalHead(head_dim_in, action_info)
+        self.value_head = FCScalarHead(head_dim_in, squeeze_scalar_dim=True)
+
+    @override
+    def forward(self, latent: Tensor) -> tuple[Distribution, Tensor]:
+        """Generate policy distributions and value estimates from latent
+        representations.
+
+        Args:
+            latent: Latent representation tensor. Shape is (*batch, len, latent_dim)
+                for sequential inputs or (*batch, latent_dim) for single-step.
 
         Returns:
             A tuple containing:
-                - Distribution representing predicted observations.
-                - Value estimates from the value head.
-                - Latent representation from the encoder.
-                - Next hidden state for the predictor.
-                - Next hidden state for the encoder.
+                - Distribution representing the policy (action probabilities).
+                - Tensor containing the estimated state value.
         """
-        return self.forward(
-            obs, upper_action, hidden_encoder, hidden_predictor, no_len=True
-        )
+        x = self.input_proj(latent)
+        x = self.core_model(x)
+        return self.policy_head(x), self.value_head(x)
 
-
-class ObsUpperActionFlattenHead(nn.Module):
-    """Head that processes observations and actions into latent
-    representations.
-
-    This head combines observation and action inputs, processes them
-    through a stacked hidden state model, and outputs latent
-    representations suitable for forward dynamics prediction.
-    """
-
-    @override
-    def __init__(
-        self,
-        obs_info: ObsInfo,
-        action_dim: int,
-        output_dim: int,
-    ) -> None:
-        """Initialize the observation-action head.
-
-        Args:
-            obs_info: Configuration for observation processing.
-            action_dim: Dimension of the action input.
-            output_dim: Dimension of the output latent representation.
-        """
-        super().__init__()
-        self.action_dim = action_dim
-        self.obs_flatten = LerpStackedFeatures(
-            obs_info.dim, obs_info.dim_hidden, obs_info.num_tokens
-        )
-        self.obs_action_proj = nn.Linear(
-            obs_info.dim_hidden + action_dim,
-            output_dim,
-        )
-
-    @override
-    def forward(self, obs: Tensor, upper_action: Tensor | None) -> Tensor:
-        """Process observations and actions into latent representations.
-
-        Args:
-            obs: Observation tensor. Shape is (*batch, len, num_tokens, obs_dim).
-            action: Action tensor. Shape is (*batch, len, action_choices).
-
-        Returns:
-            Latent representation tensor of shape (*batch, len, output_dim).
-        """
-        obs = self.obs_flatten(obs)
-        if upper_action is None:
-            upper_action = torch.zeros(
-                (*obs.shape[:-1], self.action_dim), dtype=obs.dtype, device=obs.device
-            )
-        return self.obs_action_proj(torch.cat((obs, upper_action), dim=-1))
-
-    @override
-    def __call__(self, obs: Tensor, action: Tensor) -> Tensor:
-        """Override __call__ with proper type annotations.
-
-        See forward() method for full documentation.
-        """
-        return super().__call__(obs, action)
-
-
-class LatentObsUpperActionFlattenHead(nn.Module):
-    """Head that processes observations and actions into latent
-    representations.
-
-    This head combines observation and action inputs, processes them
-    through a stacked hidden state model, and outputs latent
-    representations suitable for forward dynamics prediction.
-    """
-
-    @override
-    def __init__(
-        self,
-        obs_dim: int,
-        action_dim: int,
-        output_dim: int,
-    ) -> None:
-        """Initialize the observation-action head.
-
-        Args:
-            obs_info: Configuration for observation processing.
-            action_dim: Dimension of the action input.
-            output_dim: Dimension of the output latent representation.
-        """
-        super().__init__()
-        self.action_dim = action_dim
-        self.obs_action_proj = nn.Linear(
-            obs_dim + action_dim,
-            output_dim,
-        )
-
-    @override
-    def forward(self, obs: Tensor, upper_action: Tensor | None) -> Tensor:
-        """Process observations and actions into latent representations.
-
-        Args:
-            obs: Observation tensor. Shape is (*batch, len, obs_dim).
-            action: Action tensor. Shape is (*batch, len, action_dim).
-
-        Returns:
-            Latent representation tensor of shape (*batch, len, output_dim).
-        """
-        if upper_action is None:
-            upper_action = torch.zeros(
-                (*obs.shape[:-1], self.action_dim), dtype=obs.dtype, device=obs.device
-            )
-        return self.obs_action_proj(torch.cat((obs, upper_action), dim=-1))
-
-    @override
-    def __call__(self, obs: Tensor, action: Tensor) -> Tensor:
-        """Override __call__ with proper type annotations.
-
-        See forward() method for full documentation.
-        """
-        return super().__call__(obs, action)
-
-
-def create_hierarchical(
-    num_hierarchical_layers: int,
-    encoder: StackedHiddenState,
-    predictor: StackedHiddenState,
-    bottom_obs_upper_action_flatten_head: nn.Module,
-    latent_obs_upper_action_flatten_head: nn.Module,
-    bottom_action_dist_head: nn.Module,
-    latent_action_dist_head: nn.Module,
-    value_head: nn.Module,
-    device: torch.device | None = None,
-    dtype: torch.dtype | None = None,
-) -> list[TorchTrainingModel[LatentPolicy]]:
-    models = []
-    for i in range(num_hierarchical_layers):
-        action_dist_head = (
-            bottom_action_dist_head if i == 0 else latent_action_dist_head
-        )
-        obs_upper_action_flatten_head = (
-            bottom_obs_upper_action_flatten_head
-            if i == 0
-            else latent_obs_upper_action_flatten_head
-        )
-
-        model = LatentPolicy(
-            obs_upper_action_flatten_head=obs_upper_action_flatten_head,
-            encoder=encoder,
-            predictor=predictor,
-            action_dist_head=action_dist_head,
-            value_head=value_head,
-        )
-        models.append(
-            TorchTrainingModel(
-                model,
-                has_inference_model=True,
-                inference_procedure=LatentPolicy.forward_with_no_len,
-                device=device,
-                dtype=dtype,
-            )
-        )
-    return models
+    __call__: Callable[[Tensor], tuple[Distribution, Tensor]]
